more theatre
I had to use help to figure out that I could carry the corpse and feed that to
the monster - and I also needed help to figure out I could break the glue
bottle to stop mannequins. I *guess* the clue on the bottle sort of makes
sense, at least in hindsight, but I would never have gotten it by myself.