completed Fish!
had to look up some hints near the end. I would probably never have
gotten the cylinder and plastic bag puzzle by myself, even though it
does appear to make sense and to be properly clued. I just wasn't
thinking of the bag as something that could catch the gas from the
cylinder.

Also, had to use a hint for the final command in the game, which was
anti-climatic. I had the right idea, I could just not guess the verb
("attach").

Anyway, super fun game!